Move engine-only tests
