+++ presto-iceberg/src/test/java/io/prestosql/plugin/iceberg/TestIcebergDistributed.java
@@ -52,18 +52,6 @@ public class TestIcebergDistributed
     }
 
     @Override
-    public void testDescribeOutput()
-    {
-        // Iceberg does not support parameterized varchar
-    }
-
-    @Override
-    public void testDescribeOutputNamedAndUnnamed()
-    {
-        // Iceberg does not support parameterized varchar
-    }
-
-    @Override
     public void testCommentTable()
     {
         // Iceberg connector does not support comment on table
