+++ presto-tests/pom.xml
@@ -39,6 +39,11 @@
 
         <dependency>
             <groupId>io.prestosql</groupId>
+            <artifactId>presto-parser</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>io.prestosql</groupId>
             <artifactId>presto-testing</artifactId>
         </dependency>
